# Client, Scheduler and Worker use the new Object Manager server