--- a/scaler/__init__.py
+++ b/scaler/__init__.py
@@ -1,9 +1,9 @@
+from scaler.scheduler.scheduler_process import Scheduler
 from .about import __version__
 from .client.client import Client, ScalerFuture
 from .client.serializer.mixins import Serializer
 from .cluster.cluster import Cluster
 from .cluster.combo import SchedulerClusterCombo
-from .cluster.scheduler import Scheduler
 
 assert isinstance(__version__, str)
 assert isinstance(Client, type)