--- a/scaler/client/agent/client_agent.py
+++ b/scaler/client/agent/client_agent.py
@@ -1,6 +1,7 @@
 import asyncio
 import logging
 import threading
+from concurrent.futures import Future
 from typing import Optional
 
 import zmq.asyncio
@@ -12,6 +13,7 @@ from scaler.client.agent.object_manager import ClientObjectManager
 from scaler.client.agent.task_manager import ClientTaskManager
 from scaler.client.serializer.mixins import Serializer
 from scaler.io.async_connector import AsyncConnector
+from scaler.protocol.python.common import ObjectStorageAddress
 from scaler.protocol.python.message import (
     ClientDisconnect,
     ClientHeartbeatEcho,
@@ -19,7 +21,6 @@ from scaler.protocol.python.message import (
     GraphTask,
     GraphTaskCancel,
     ObjectInstruction,
-    ObjectRequest,
     Task,
     TaskCancel,
     TaskResult,
@@ -34,7 +35,7 @@ class ClientAgent(threading.Thread):
     def __init__(
         self,
         identity: bytes,
-        client_agent_address: ZMQConfig,
+        agent_address: ZMQConfig,
         scheduler_address: ZMQConfig,
         context: zmq.Context,
         future_manager: ClientFutureManager,
@@ -51,9 +52,10 @@ class ClientAgent(threading.Thread):
         self._serializer = serializer
 
         self._identity = identity
-        self._client_agent_address = client_agent_address
+        self._agent_address = agent_address
         self._scheduler_address = scheduler_address
         self._context = context
+        self._storage_address: Future[ObjectStorageAddress] = Future()
 
         self._future_manager = future_manager
 
@@ -62,7 +64,7 @@ class ClientAgent(threading.Thread):
             name="client_agent_internal",
             socket_type=zmq.PAIR,
             bind_or_connect="bind",
-            address=self._client_agent_address,
+            address=self._agent_address,
             callback=self.__on_receive_from_client,
             identity=None,
         )
@@ -83,22 +85,22 @@ class ClientAgent(threading.Thread):
     def __initialize(self):
         self._disconnect_manager = ClientDisconnectManager()
         self._heartbeat_manager = ClientHeartbeatManager(death_timeout_seconds=self._timeout_seconds)
-        self._object_manager = ClientObjectManager()
+        self._object_manager = ClientObjectManager(identity=self._identity)
         self._task_manager = ClientTaskManager()
 
         # register all managers
         self._disconnect_manager.register(
             connector_internal=self._connector_internal, connector_external=self._connector_external
         )
-        self._object_manager.register(connector_internal=self._connector_internal, future_manager=self._future_manager)
+        self._object_manager.register(
+            connector_internal=self._connector_internal, connector_external=self._connector_external
+        )
         self._task_manager.register(
             connector_external=self._connector_external,
             object_manager=self._object_manager,
             future_manager=self._future_manager,
         )
-        self._heartbeat_manager.register(
-            connector_external=self._connector_external, object_manager=self._object_manager
-        )
+        self._heartbeat_manager.register(connector_external=self._connector_external)
 
     def __run_loop(self):
         self._loop = asyncio.new_event_loop()
@@ -110,8 +112,9 @@ class ClientAgent(threading.Thread):
         self.__initialize()
         self.__run_loop()
 
-    def wait_until_ready(self):
-        self._object_manager.wait_until_ready()
+    def get_storage_address(self) -> ObjectStorageAddress:
+        """Returns the object storage configuration, or block until it receives it."""
+        return self._heartbeat_manager.get_storage_address()
 
     async def __on_receive_from_client(self, message: Message):
         if isinstance(message, ClientDisconnect):
@@ -177,6 +180,9 @@ class ClientAgent(threading.Thread):
 
         if exception is None:
             return
+
+        if not self._storage_address.done():
+            self._storage_address.set_exception(exception)
 
         if isinstance(exception, asyncio.CancelledError):
             logging.error("ClientAgent: async. loop cancelled")