--- a/scaler/client/agent/future_manager.py
+++ b/scaler/client/agent/future_manager.py
@@ -6,9 +6,8 @@ from typing import Dict, Tuple
 from scaler.client.agent.mixins import FutureManager
 from scaler.client.future import ScalerFuture
 from scaler.client.serializer.mixins import Serializer
-from scaler.io.utility import concat_list_of_bytes
 from scaler.protocol.python.common import TaskStatus
-from scaler.protocol.python.message import ObjectResponse, TaskCancel, TaskResult
+from scaler.protocol.python.message import TaskCancel, TaskResult
 from scaler.utility.exceptions import DisconnectedError, NoWorkerError, TaskNotFoundError, WorkerDiedError
 from scaler.utility.metadata.profile_result import retrieve_profiling_result_from_task_result
 from scaler.utility.object_utility import deserialize_failure
@@ -100,22 +99,17 @@ class ClientFutureManager(FutureManager):
         with self._lock:
             self._task_id_to_future.pop(task_cancel.task_id, None)
 
-    def on_object_response(self, response: ObjectResponse):
-        for object_id, object_name, object_bytes in zip(
-            response.object_content.object_ids,
-            response.object_content.object_names,
-            response.object_content.object_bytes,
-        ):
-            if object_id not in self._object_id_to_future:
-                continue
+    def on_object_storage_get_response(self, object_id: bytes, payload: bytes):
+        if object_id not in self._object_id_to_future:
+            return
 
-            status, future = self._object_id_to_future.pop(object_id)
+        status, future = self._object_id_to_future.pop(object_id)
 
-            try:
-                if status == TaskStatus.Success:
-                    future.set_result(self._serializer.deserialize(concat_list_of_bytes(object_bytes)))
+        try:
+            if status == TaskStatus.Success:
+                future.set_result(self._serializer.deserialize(payload))
 
-                elif status == TaskStatus.Failed:
-                    future.set_exception(deserialize_failure(concat_list_of_bytes(object_bytes)))
-            except InvalidStateError:
-                continue  # future got canceled
+            elif status == TaskStatus.Failed:
+                future.set_exception(deserialize_failure(payload))
+        except InvalidStateError:
+            return  # future got canceled