‼️ **Work In Progress** ‼️

This PR removes the object's data management from the scheduler, moving it to the new C++ Object Manager.

Here are is a diagram with the new data flows:

![scaler_object_storage](https://github.com/user-attachments/assets/a202406b-7799-41a1-93db-77be58a5740a)

Here are the major changes:

- Two new network connectors that link directly to the object storage server (`AsyncObjectStorageConnector` and `SyncObjectStorageConnector`);
- The client and its futures directly connect to the object storage. Object's data do not get transmitted to the client's agent, only the `ObjectInstruction` messages do;
- Only the scheduler sends `Delete` messages to the storage server, except for the task result objects, that are also immediately deleted by the client when received. The scheduler might `Delete` these object a second time, but that's OK;
- There is an unexpected async connector that runs inside the worker's agent. Its only purpose is to create the exception objects when a processor fails;
- The Symphony worker has been greatly simplified, as the async object connector removes the need for the `asyncio` lock logic from the previous implementation @1597463007.