@sharpener6 2025-05-14
@rafa-be @gxuu ,
There is one edge case that current object storage server protocol will not work:
That is for test case `TestClient.test_clear`
so if user use `send_object` api which actually send object to object storage, but then quickly deleted the object, so if we send a task that uses this staled object reference, before the change, it will raise MissingObject exception and propagate to the frontend. but since now object storage will block until it have data in the object storage server, this will cause this task hanging forever