@gxuu 2025-05-15
> if user use `send_object` api which actually send object to object storage, but then quickly deleted the object
Should the user ever do that? And should we support this?

There isn't much to do in the server side. 
- Suppose server remember whether an object id appeared before in the system, when a new request comes in, there is no way for the server side to distinguish whether it should hang or sending "once appeared in the system but now removed" response).
- Suppose server adding timeout mechanism and postpone all delete request, this will still not work because how long should we set the timeout to be? 

The ideal way to do this is to somehow make the scheduler get request reach sooner than delete object request, which can be done if we use the same tcp connection.