@rafa-be 2025-05-15
> @rafa-be @gxuu , There is one edge case that current object storage server protocol will not work: That is for test case `TestClient.test_clear` so if user use `send_object` api which actually send object to object storage, but then quickly deleted the object, so if we send a task that uses this staled object reference, before the change, it will raise MissingObject exception and propagate to the frontend. but since now object storage will block until it have data in the object storage server, this will cause this task hanging forever

It should be OK with the current version of the Scheduler's state-machine, as clearing the client's objects will cancel all the tasks first:

```python
class Client:
     ...
     def clear(self):
        self._future_manager.cancel_all_futures()
        self._object_buffer.clear()
```

Thus the `TaskCancel(force=True)` will `SIGTERM` the processor process that is waiting on the `GET` request. Once cleared, you cannot use the old object references for new tasks, as these got invalidated.

If we remove this `force=True` feature, I think we could somehow fix the problem by making the scheduler delay the `DELETE` requests for objects that are associated with a running task.

>   Suppose server adding timeout mechanism and postpone all delete request, this will still not work because how long should we set the timeout to be?

That might still be desirable, as a kind of safety trigger, and to easily detect these problems. But it does not have to be implemented on the server side.
 
> The ideal way to do this is to somehow make the scheduler get request reach sooner than delete object request, which can be done if we use the same tcp connection.

Unfortunately, these don't use the same TCP link, as the scheduler makes the `DELETE` requests, while the worker makes the `GET` requests. But I can prevent the scheduler to send the `DELETE` request if it's associated task is running.