@rafa-be 2025-05-29
I pushed a major update to this PR.

**All unit tests are now passing successfully**.

The general architecture hasn't changed (i.e. the diagram above is still valid).

There are the major changes:

- the **worker and processor initialization have been re-written**, that's because of the connection to the object storage server that can only be made once the connection to the scheduler has been made;
- the **object IDs are now randomly generated** (instead of based on the md5sum of their content). That's is required because there might be synchronization issues between the client, scheduler and object storage server when a client do a `clear()` request;
- I **fixed a few issues in the `ScalerFuture` class**, related to how and when result objects are obtained from the storage server;
- all **IDs (Task ID, Object ID, Worker ID, Worker ID...) now have their respective types**, instead of the previous `bytes`type.
    - that helps avoiding mistakes when dealing with these values;
    - IDs were generated by calling the `uuid` module directly, sometimes inconsistently, now it's "standardized" in these classes.  
- the **C++ object storage server got wrapped into an `ObjectStorageServer` Python class**.
    - API is now similar to the one of `Scheduler` and `Worker`;
    - the server exposes an event file-descriptor that allows the Python code to be notified when the server is ready (i.e. when the TCP socket is open).