@rafa-be 2025-05-30
> Great work!
> 
> There are a few things that I'm in doubt with, they are not necessarily associated with this PR though:
> 
>     * I was inspecting the `client_manager.py`'s code. I found that `client` (which in your PR is renamed to `client_id` ) is quite large (IIRC 46 Bytes). This is way longer than `task_id` which is only 16 bytes. If they are all IDs then why length differ so much?
> 
>     * Is now a good time to fix the length of various kinds of IDs? Fixed size ID's are much easier to deal with in C++ code, and is easier to bound check.
> 
> 
> And thanks for the improvement of object storage server! It was changing so drastically within a tight schedule, so I postpone lots of clean up.

I haven't changed how client, task and worker IDs are generated, so these keep the old format, which can basically be any size as it's an UTF-8 encoded string.
Object IDs have been changed to the 32-bytes format used by the object storage server.

But I agree, it would be nice to standardize all these IDs. What I was thinking was to have random 16-bytes IDs for worker and clients, and 32-bytes IDs for tasks and objects, with the first 16-bytes always referring to the owner of the object (i.e. the client).

For example, tasks and object IDS would always like this:

```
+--------------+-----------------------+
|   16-bytes   |       16-bytes        |
+--------------+-----------------------+
| Client ID    | Random task/object ID |
+--------------+-----------------------+
```

That would help debugging whom a task or object comes from.