@rafa-be 2025-05-30
@sharpener6 Here are the benchmarks of the new object storage against the old one on some relevant unit tests we have:

| Test name                             | Old Real Time | Old User Time | New Real Time     | New User Time     |
|--------------------------------------|----------------|----------------|--------------------|--------------------|
| `TestClient.test_noop_submit`        | 13.061s        | 21.504s        | 14.567s (+12%)     | 27.949s (+30%)     |
| `TestClient.test_noop_map`           | 9.725s         | 20.125s        | 11.959s (+23%)     | 22.202s (+10%)     |
| `TestClient.test_heavy_function`     | 13.178s        | 23.425s        | 15.835s (+20%)     | 32.524s (+39%)     |
| `TestClient.test_very_large_payload` | 7.484s         | 3.984s         | 4.130s (−45%)      | 2.038s (−49%)      |
| `TestGraph.test_graph`               | 3.523s         | 1.620s         | 1.827s (−48%)      | 1.401s (−14%)      |

As expected, the new storage is faster on large payload, but slower on small tasks. That's kind of expected, as the client uses a synchronized connector to the storage server, which adds a little bit of latency.